Add macro TAI_NEXT to continue hook chain
Instead of taking only the return type and casting cur->old and
next->func to variadic function pointers, we take the enclosing
function and use the GCC extension "typeof" to cast to the exact
function pointer type. This grants two advantages:

1. Static type checking on the hook continuation call.
2. Prevents double promotion of single precision float arguments, which
is required by the C standard on calls to variadic functions.